Обновление документации API, добавление интеграционных тестов, добавление задачи для генерации документации API

+++ core/src/main/resources/db/migration/V1__initialize.sql
@@ -1,7 +1,7 @@
 create table if not exists bill (
     id uuid DEFAULT gen_random_uuid(),
     user_id uuid not null,
-    balance decimal(19, 2) not null DEFAULT 0,
+    balance bigint not null DEFAULT 0,
     creation_instant timestamptz not null DEFAULT now(),
     closing_instant timestamptz,
     primary key (id)
@@ -10,7 +10,7 @@ create table if not exists bill (
 create table if not exists transaction (
     id uuid DEFAULT gen_random_uuid(),
     bill_id uuid not null,
-    balance_change decimal(19, 2) not null,
+    balance_change bigint not null,
     reason varchar not null,
     instant timestamptz not null DEFAULT now(),
     primary key (id),
